feat: Introduce Custom Bar Overlays
